Support searching flights for multiple dates and destinations
Tries different combinations of dates for both direct and return flights in order to find otions with the lowest fare. Results are merged, groupped and saved in a file per trip direction. The desired trips can be configured in application.conf where it is possible to specify airports, dates and trip duration. For example, if I want to travel to IEV from SYD between December 17 and December 23 for 26-30 days, the config will look like this:

tripConfigs = [
    {
      "fromAirport": [
        "SYD"
      ],
      "toAirport": [
        "IEV"
      ],
      "minDate": "17.12.2016",
      "maxDate": "23.12.2016",
      "minDuration": 26,
      "maxDuration": 30,
    }
  ]

It is possible to specify multiple destinations, for example I can say "I want to travel from SYD to IEV or LHR, depending what is cheaper". Also, it is possible to configure multiple trips.